Added separate command for search results Available links: -/teams/search - form of searching team -/teams/search/results?name={name} - search results by name parameter -/teams/details?id={id} - detailed information about team by id parameter
