+++ src/main/java/com/andersen/corgisteam/corgiroulette/servlet/command/impl/SearchTeamFormCommand.java
@@ -1,0 +1,18 @@
+package com.andersen.corgisteam.corgiroulette.servlet.command.impl;
+
+import com.andersen.corgisteam.corgiroulette.servlet.command.Command;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class SearchTeamFormCommand implements Command {
+
+    public static final String SEARCH_TEAM_PATH = "/WEB-INF/jsp/team/searchTeam.jsp";
+
+    @Override
+    public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher(SEARCH_TEAM_PATH).forward(request, response);
+    }
+}
