+++ src/main/java/com/andersen/corgisteam/corgiroulette/servlet/command/impl/TeamSearchResultsCommand.java
@@ -3,7 +3,6 @@ package com.andersen.corgisteam.corgiroulette.servlet.command.impl;
 import com.andersen.corgisteam.corgiroulette.entity.Team;
 import com.andersen.corgisteam.corgiroulette.repository.QueryExecutionException;
 import com.andersen.corgisteam.corgiroulette.service.TeamService;
-import com.andersen.corgisteam.corgiroulette.service.exception.RequiredFieldIsEmptyException;
 import com.andersen.corgisteam.corgiroulette.service.exception.ValidationException;
 import com.andersen.corgisteam.corgiroulette.servlet.command.Command;
 
@@ -11,30 +10,30 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 
-public class TeamDetailsCommand implements Command {
+public class TeamSearchResultsCommand implements Command {
 
     private TeamService teamService;
 
-    private static final String TEAM_DETAILS_PATH = "/WEB-INF/jsp/team/teamDetails.jsp";
+    private static final String TEAM_SEARCH_RESULTS = "/WEB-INF/jsp/team/teamSearchResults.jsp";
 
-    private static final String ID_PARAMETER = "id";
-    private static final String TEAM_PARAMETER = "team";
+    private static final String NAME_PARAMETER = "name";
+    private static final String TEAMS_PARAMETER = "teams";
     private static final String ERROR_ATTRIBUTE_NAME = "errorMessage";
 
-    public TeamDetailsCommand(TeamService teamService) {
+    public TeamSearchResultsCommand(TeamService teamService) {
         this.teamService = teamService;
     }
 
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        try{
-            long teamId = Long.parseLong(request.getParameter(ID_PARAMETER));
-            Team team = teamService.get(teamId);
-
-            request.setAttribute(TEAM_PARAMETER, team);
-            request.getRequestDispatcher(TEAM_DETAILS_PATH).forward(request, response);
-        } catch (QueryExecutionException e){
+        try {
+            String name = request.getParameter(NAME_PARAMETER);
+            List<Team> teams = teamService.getAllByName(name);
+            request.setAttribute(TEAMS_PARAMETER, teams);
+            request.getRequestDispatcher(TEAM_SEARCH_RESULTS).forward(request, response);
+        } catch (QueryExecutionException | ValidationException e) {
             request.setAttribute(ERROR_ATTRIBUTE_NAME, e.getMessage());
             request.getRequestDispatcher(SearchTeamFormCommand.SEARCH_TEAM_PATH).forward(request, response);
         }
