+++ src/main/webapp/WEB-INF/jsp/team/teamDetails.jsp
@@ -11,9 +11,9 @@
     <p>Id: ${team.id}</p>
     <p>Name: ${team.name}</p>
 
-    <p><a href="${pageContext.request.contextPath}/teams/edit?id=${team.id}">Edit</a></p>
-    <p><a href="${pageContext.request.contextPath}/teams/delete?id=${team.id}">Delete</a></p>
-    <p><a href="${pageContext.request.contextPath}/teams/search">Use search</a></p>
-    <p><a href="${pageContext.request.contextPath}/teams">Show all teams</a></p>
+    <p><a href="${pageContext.request.contextPath}/roulette/teams/edit?id=${team.id}">Edit</a></p>
+    <p><a href="${pageContext.request.contextPath}/roulette/teams/delete?id=${team.id}">Delete</a></p>
+    <p><a href="${pageContext.request.contextPath}/roulette/teams/search">Use search</a></p>
+    <p><a href="${pageContext.request.contextPath}/roulette/teams">Show all teams</a></p>
 </body>
 </html>
